requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
djangorestframework==3.14.0
frozenlist==1.3.3
idna==3.4
lxml==4.9.2
multidict==6.0.4
pytz==2023.3
requests==2.28.2
soupsieve==2.4
sqlparse==0.4.3
urllib3==1.26.15
yarl==1.8.2
